requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dependency-injector-fork==4.42.1
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
orjson==3.9.10
pydantic==2.5.3
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.17.2
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
setuptools==69.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
typer==0.9.0
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.42.0
debugpy==1.8.1
jinja2